# Conflicts:
#	target/m2e-wtp/web-resources/META-INF/maven/com.store/greenStore/pom.properties
#	target/test-classes/test/TestMain.class

+++ src/main/webapp/WEB-INF/views/notice/listAll.jsp
@@ -1,5 +1,4 @@
-<%@ page language="java" contentType="text/html; charset=UTF-8"
-	pageEncoding="UTF-8"%>
+<%@ page language="java" contentType="text/html; charset=UTF-8" pageEncoding="UTF-8"%>
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt"%>
 <%@ taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
@@ -30,12 +29,14 @@
           <div class="box box-solid">
             <div class="box-header with-border">
               <h2 class="box-title" style="margin:20px;">GreenStore 공지사항을 알려드립니다!</h2>
-              
-            
+              <c:if test="${not empty member }"> 
+              <c:if test="${ member.mname == admin }"> 
+              	<a href="http://localhost:8080/greenStore/notice/write" class="btn">글쓰기</a>
+			  	</c:if> 
+			  </c:if> 
             </div>
             <c:forEach var="notice" items="${ list }">
 	            <div class="box-body">
-	            
 	              <div class="box-group" id="accordion">
 	                <div class="panel box box-success">
 	                  <div class="box-header with-border">
@@ -52,30 +53,15 @@
 	                    <div class="box-body">
 							${ notice.ncontent }
 	                    </div>
-							<%-- <div class="box-body" style="position:right;">
-								<a href="update?nkey=${ notice.nkey }" class="btn btn-warning">글수정하기</a>
-							 	<!-- <button type="submit" class="btn btn-danger">글삭제하기</button> -->
-							 	<a href="delete?nkey=${ notice.nkey }" data-confirm="삭제하시겠습니까?" class="btn btn-danger">글삭제하기</a>
-								<button type="submit" class="btn btn-primary">전체글보기</button>
-	                    	</div>  --%>
 	                    		<div class="pull-right">
-	                    		<form role="form" method="post">
-									<input type="hidden" name="nkey" value="${notice.nkey}">
-								
-                		 <a href="http://localhost:8080/greenStore/notice/write" class="btn">글쓰기</a>
-								<a
-									href="update?nkey=${notice.nkey}"
-									class="btn"> <i class="icon-list"></i> 수정
-						
-								</a> <a
-									href="delete?nkey=${notice.nkey}"
-									class="btn" data-confirm="삭제하시겠습니까?"> <i class="icon-remove"></i>
-									삭제
-								</a> <a href="listAll"
-									class="btn"> <i class="icon-list"></i> 목록으로
-								</a> 
+	                    		<c:if test="${not empty member }"> 
+	                    		<c:if test="${ member.mname == admin }">
+									<a href="update?nkey=${notice.nkey}" class="btn" id="btn-primary"> <i class="icon-list"></i> 수정</a>
+									<a href="delete?nkey=${notice.nkey}" class="btn" data-confirm="삭제하시겠습니까?"> <i class="icon-remove"></i>삭제</a>
+								 	</c:if>
+								</c:if> 
+								<a href="listAll" class="btn"> <i class="icon-list"></i> 목록으로</a> 
 							</div>
-							</form>
 	                  </div>
 	                </div>
 	              </div>
@@ -85,5 +71,4 @@
         </div>
 	</div>
 </div>
-
